# apps/api/src/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Telecom FieldOps API
  version: 1.0.0
  description: Security-focused API for authentication, user management, roles and audit-ready operations.
servers:
  - url: http://localhost:3000/api/v1
...
components:
  securitySchemes:
    OAuth2Password:
      type: oauth2
      flows:
        password:
          tokenUrl: /api/v1/auth/login
          scopes: {}
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, instance, correlationId]
      properties:
        type:
          type: string
          example: urn:telecom:error:validation
        title:
          type: string
          example: Validation Error
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request body validation failed.
        instance:
          type: string
          example: /api/v1/users
        correlationId:
          type: string
          example: c123
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    UserPublic:
      type: object
      required: [id, email, blocked, roles]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        blocked:
          type: boolean
        roles:
          type: array
          items:
            type: string
...
    CreateUserRequest:
      type: object
      required: [email, password, roles]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        roles:
          type: array
          minItems: 1
          items:
            type: string
    UpdateUserRequest:
      type: object
      minProperties: 1
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        blocked:
          type: boolean
        roles:
          type: array
          minItems: 1
          items:
            type: string
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /users/{id}:
    patch:
      tags: [Users]
      summary: Update user data and roles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}/block:
    post:
      tags: [Users]
      summary: Block user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublic'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...